Update: Testes Account

+++ src/test/java/qbank/sistemabancario/context/conta/contaTest/ContaCorrentTest.java
@@ -4,18 +4,20 @@ import org.junit.jupiter.api.Test;
 import qbank.sistemabancario.context.conta.ContaCorrente;
 import qbank.sistemabancario.context.pessoa.Pessoa;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ContaCorrentTest {
 
     Pessoa pessoatest = new Pessoa("Rafael","00100100100","000111000","rafael@gmail","31123456789");
-    ContaCorrente conta = new ContaCorrente("001","123",50000F,pessoatest);
+    ContaCorrente conta = new ContaCorrente("001","123",50000f,pessoatest);
 
     @Test
     void RetornaSaldoDaContaCorrente() {
-        float saldoConta = conta.getSaldo();
+        BigDecimal saldoConta = conta.getSaldo();
 
-        assertEquals(saldoConta,50000F);
+        assertEquals(saldoConta, BigDecimal.valueOf(50000f));
     }
     @Test
     void RetornaAgenciDaContaCorrente() {
@@ -29,8 +31,8 @@ public class ContaCorrentTest {
     void RetornaDadosPessoais() { assertEquals(conta.getPessoa(),pessoatest); }
     @Test
     void AlterandoSaldoContaCorrente() {
-        conta.setSaldo(100.50F);
-        assertEquals(conta.getSaldo(),100.50F);
+        conta.setSaldo(BigDecimal.valueOf(100.50F));
+        assertEquals(conta.getSaldo(), BigDecimal.valueOf(100.50F));
     }
     @Test
     void AlterandoAgenciaContaCorrente() {
